+++ src/Hooks/PlayerObject.hpp
@@ -6,6 +6,12 @@
 
 class $modify(ProPlayerObject, PlayerObject) {
 
+	static void onModify(auto& self) {
+        (void)self.setHookPriorityPre("PlayerObject::playDeathEffect", Priority::Last + 0x500000); // eclipses big ass priority didnt test other mod menus
+        (void)self.setHookPriorityPost("PlayerObject::playSpawnEffect", Priority::Last);
+        (void)self.setHookPriorityPost("PlayerObject::playCompleteEffect", Priority::Last);
+	}
+
 	struct Fields {
 		IconType m_previousType = static_cast<IconType>(-9038);
 
@@ -14,10 +20,14 @@ class $modify(ProPlayerObject, PlayerObject) {
 		CCSprite* m_vehicleSprite = nullptr;
 		CCSprite* m_vehicleSpriteSecondary = nullptr;
 
+		std::vector<CCSprite*> m_animSprites;
+
 		bool m_thatOneUfoShipAndCubeModIsLoaded = false;
 	};
 
 	IconType getIconType();
+
+	void updateVisibility(bool = false);
 
 	void updateSprite(CCSprite*, CCSprite*&, SpriteType, bool);
 
@@ -50,5 +60,11 @@ class $modify(ProPlayerObject, PlayerObject) {
 	void createSpider(int);
 
 	bool init(int, int, GJBaseGameLayer*, cocos2d::CCLayer*, bool);
+
+	void playCompleteEffect(bool, bool);
+
+	void playDeathEffect();
+
+	void playSpawnEffect();
 
 };