+++ changelog.md
@@ -1,3 +1,8 @@
+# 1.0.1
+
+* Fixed icon being white when gradient was empty.
+* Fixed gradient not disappearing when dying or completing level.
+
 # 1.0.0
 
 * firt